Initial import of basic inline markup and structural parsing for bills
This parses the XML format Congress uses for publishing bills. It is not
yet complete but is able to capture the outline of the bill and the
text portions of each structural element.

In particular it currently lacks support for the block elements that are
permitted between the text and continuation text in each structural
element.