Block markup elements
The final major class of element in the bills XML schema is block
elements. These appear as direct children of structural elements, between
the text (if any) and the child structural elements.

These are the most variously-structured of the three classes, with each
type having its own entirely-different internal structure. Any caller
working with these will need a separate implementation for each type,
because they have nothing useful in common.