(Incomplete) utility for walking the parsed DOM
When converting the XML format into another format it will be common to
traverse the DOM in the order in which the different elements/nodes will
be rendered. The Walk functions provide this, using caller-provided
visitor implementations to specify behavior for each node.

The primary use-case here is serializing to other formats, but it can
also be used to, for example, extract an outline of the captioned
structural elements and discard everything else.